requirements.txt
fastmcp
httpx
python-dotenv
loguru
